Add files via upload
A sample presentation from one of our biweekly meetings. The goal was to create a script that is able to calculate the correlation of gene expressions for pairs of points in the control tissue sample, as a function of varying distances. The challenge of this was computational time and memory (we have to find all possible point-pairs for each distance). So, a major focus of the presentation was not just discussing conclusions and unexpected findings, but also overviewing the logic behind the algorithm, identifying key assumptions and confirming them.