Add files via upload
Sample workflow for 2D correlation analysis. Goal was that, for our pre-selected gene sets, we are able to obtain 2D correlation data (in both x and y direction) as a function of distance and output results as a heatmap. I did this for both wound-boundary (section 1) and control (section 2) data. Note that the function sweep2D is a modified version of the original sweep1D function, except that distance the scalar is now distance the vector.